fix(protocol): correct 5 byte-level issues in 15.11 login packets
- Fix challenge packet (0x1F): skipBytes(-13) -> (-12), checksum 9 -> 8
- Fix sendIcons (0xA2): icons field from U32 to U64 (client expects 8 bytes)
- Fix AddPlayerSkills (0xA1): addDouble precision from 2/3 to 4
- Fix sendWorldLight (0x82): add isAccessPlayer check for GM brightness
- Fix sendPremiumTrigger (0x9E): send 16 triggers instead of empty list

+++ src/player.cpp
@@ -393,9 +393,9 @@ float Player::getDefenseFactor() const
 	}
 }
 
-uint32_t Player::getClientIcons() const
+uint64_t Player::getClientIcons() const
 {
-	uint32_t icons = 0;
+	uint64_t icons = 0;
 	for (Condition* condition : conditions) {
 		if (!isSuppress(condition->getType())) {
 			icons |= condition->getIcons();
